% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calculate.object.digest}
\alias{calculate.object.digest}
\title{Calculates object's digest}
\usage{
calculate.object.digest(objectname, target.environment = NULL,
  flag_use_attrib = TRUE, flag_add_attrib = FALSE)
}
\arguments{
\item{object}{The name of the object to calculate digest of. Object must exist in
\code{.GlobalEnv}.}
}
\value{
Lowercase MD5 string with the digest of the object
}
\description{
It efficiently calculates object's digest.
}
\details{
It treats \code{data.frame} objects separately. Rather than invoking \code{digest::digest()}
on them directly, it first splits the object into individual columns, sorts them,
calculates the
digest on each column separately (trying to do that in parallel), and merges the
results into one string, that is digested again.
}
