% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locks.R
\name{create.lock.file}
\alias{create.lock.file}
\title{Creates lock file on a given path. Lock file is just a 0-length file.}
\usage{
create.lock.file(path, timeout = NULL)
}
\arguments{
\item{path}{Path to the lock file}

\item{timeout}{Wall time after which the lock will be considered open regardless if someone opened it}
}
\description{
The file will have an extension getOption('objectstorage.lock_extension')
}
