% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{create_objectstorage}
\alias{create_objectstorage}
\title{Makes sure there is a objectstorage on the specified path.}
\usage{
create_objectstorage(storagepath)
}
\arguments{
\item{storagepath}{Path where the storage should be created}
}
\value{
Returns updated metadata on success, and \code{NULL} on failure.
}
\description{
Makes sure there is a objectstorage on the specified path.
}
