% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{get_full_digest}
\alias{get_full_digest}
\title{Returns md5 hash of all the objects in the \code{objectstorage} stored in disk.}
\usage{
get_full_digest(storagepath)
}
\arguments{
\item{storagepath}{Path with the storage.}
}
\value{
Returns data stamp of typ \code{c("POSIXct", "POSIXt")} of the modification time.
}
\description{
This function is almost instantenous even for large objects, because it doesn't
read the actual objects, it reads their hashes from the description file.
}
\details{
Actually this ability to quickly get the cryptographic hash of all objects contained within
the archive is a motivation behind
}
