% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{get_mtime}
\alias{get_mtime}
\title{Returns mtime of last modification of the \code{objectstorage}}
\usage{
get_mtime(storagepath)
}
\arguments{
\item{storagepath}{Path with the storage.}
}
\value{
Returns data stamp of typ \code{c("POSIXct", "POSIXt")} of the modification time.
}
\description{
Returns mtime of last modification of the \code{objectstorage}
}
