% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{get_object}
\alias{get_object}
\title{Shortcut function that returns a single object read from the storage}
\usage{
get_object(storagepath, objectname = NULL,
  flag_double_check_digest = FALSE)
}
\arguments{
\item{storagepath}{Path to the storage metadata}

\item{flag_double_check_digest}{Calculate the digest of the extracted objects and check it against the metadata}

\item{objectnames}{Objectnames to extract}
}
\value{
the requested object.
}
\description{
Shortcut function that returns a single object read from the storage
}
