% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{list_runtime_objects}
\alias{list_runtime_objects}
\title{Lists all the runtime objects cotained in the task with path path.}
\usage{
list_runtime_objects(storagepath)
}
\arguments{
\item{storagepath}{Path with the storage.}
}
\value{
Returns `data.frame` with the following columns:
\describe{
\item{\strong{objectname}}{Name of the stored object. This is a primary key.}
\item{\strong{digest}}{String with the digest of the object.}
\item{\strong{size}}{Numeric value with the size of the stored object.}
\item{\strong{archive_filename}}{Path where the object is stored absolute or relative to the storage path.}
\item{\strong{single_object}}{Logical. \code{TRUE} if the archive contain only this one object. Otherwise
archive contains named list of objects.}
}
}
\description{
Lists all the runtime objects cotained in the task with path path.
}
