% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R, R/runtime_objects.R
\name{load_objects}
\alias{load_objects}
\alias{load_objects}
\title{Loads specified objects into the given environment.}
\usage{
load_objects(storagepath, objectnames, target_environment,
  flag_double_check_digest = FALSE, aliasnames = NULL)

load_objects(storagepath, objectnames, target_environment,
  flag_double_check_digest = FALSE, aliasnames = NULL)
}
\arguments{
\item{storagepath}{Path to the storage}

\item{objectnames}{Vector with names of objects to load. Defaults to all objects}

\item{target_environment}{Target environment where to put the objects}

\item{flag_double_check_digest}{Calculate the digest of the extracted objects and check it against the metadata}

\item{aliasnames}{Optional vector of the same length as \code{objectnames} with new names for the extracted objects.}

\item{envir}{The target environment. Defaults to the global environment.}

\item{storagepath}{Path to the storage metadata}

\item{objectnames}{Objectnames to extract}
}
\value{
Logical vector, one for each loaded object. \code{TRUE} means that
        load was successfull, \code{FALSE} otherwise.
}
\description{
Loads specified objects into the given environment.

Main function to load the objects into the given environment
}
