% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locks.R
\name{lock.exists}
\alias{lock.exists}
\title{Tests if the lock is locked}
\usage{
lock.exists(path, timeout = NULL)
}
\arguments{
\item{path}{Path to the lock file}

\item{timeout}{Wall time after which the lock will be considered open regardless if someone opened it}
}
\description{
Tests if the lock is locked
}
