% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{modify_runtime_objects}
\alias{modify_runtime_objects}
\title{Adds or removes objects in the storage.}
\usage{
modify_runtime_objects(storagepath, obj.environment, objects_to_add = NULL,
  objects_to_remove = character(0), flag_forced_save_filenames = FALSE,
  flag_use_tmp_storage = FALSE, forced_archive_paths = NA,
  compress = "gzip", large_archive_prefix = NULL, locktimeout = NULL,
  wait_for = "save", parallel_cpus = NULL)
}
\arguments{
\item{storagepath}{Path with the storage.}

\item{obj.environment}{Environment or named list with the objects}

\item{addobjectnames}{Character vector with the names of the objects to add. Defaults to all objects in the
\code{obj.environment}.}

\item{removeobjectnames}{Character vector with the names of the objects to remove. Cannot contain objects
listed in \code{addobjectnames}.}

\item{archive_filename}{Optional character vector with custom paths to the archives. Can be a single character object,
vector with the size of \code{objects_to_add} or named vector with keys from \code{objects_to_add}.}
}
\value{
Returns `data.frame` with the following columns:
\describe{
\item{\strong{objectname}}{Name of the stored object. This is a primary key.}
\item{\strong{digest}}{String with the digest of the object.}
\item{\strong{size}}{Numeric value with the size of the stored object.}
\item{\strong{archive_filename}}{Path where the object is stored absolute or relative to the storage path.}
\item{\strong{single_object}}{Logical. \code{TRUE} if the archive contain only this one object. Otherwise
archive contains named list of objects.}
\item{\strong{compress}}{Type of compression used to store this individual object}
}
}
\description{
Adds or removes objects in the storage.
}
