% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime_objects.R
\name{parse_argument}
\alias{parse_argument}
\title{Function parses the argument generating named vector with keys objectnames}
\usage{
parse_argument(arg, objectnames, default_value)
}
\arguments{
\item{arg}{The argument to parse. Can be a single value, vector of lenght equal to
the length of \code{objectnames} or named vector with keys from \code{objectnames}.}

\item{objectnames}{Objectnames that sets the basis for the intepretation of the \code{arg}}

\item{default_value}{Value in case \code{arg} is a named vector with some \code{objectnames}
missing its values.}
}
\description{
Function parses the argument generating named vector with keys objectnames
}
