% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{remove_all}
\alias{remove_all}
\title{Removes everything from disk}
\usage{
remove_all(storagepath)
}
\arguments{
\item{storagepath}{Path to the storage}
}
\description{
Removes everything from disk
}
