% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save_large_object}
\alias{save_large_object}
\title{Saves arbitrary large object to disk using saveRDS. If compression is 'xz', then
the file is first saved quickly with no compression, and then the background
process is spawned that compresses the file in multithreaded fassion using `pxz`,
if the program is available.}
\usage{
save_large_object(obj, file, compress = "xz", wait_for = c("save",
  "compress", "none"), flag_use_tmp_storage = FALSE,
  fn_to_run_after_save = NULL, fn_to_run_after_compress = NULL,
  parallel_cpus = NULL, flag_detach = FALSE)
}
\arguments{
\item{obj}{The object to be saved}

\item{file}{Path to the file}

\item{compress}{Compression method with the same meaning as saveRDS. Default is 'xz'.}

\item{wait}{If set the function exits only after the object is available to read.}

\item{fn_to_run_afters_save}{Function that will be run after save. The function will get a single argument - path to the
newly created file}
}
\value{
parallel job with the backgroud save, if wait is not 'none'
}
\description{
Saves arbitrary large object to disk using saveRDS. If compression is 'xz', then
the file is first saved quickly with no compression, and then the background
process is spawned that compresses the file in multithreaded fassion using `pxz`,
if the program is available.
}
