% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{save_objects}
\alias{save_objects}
\title{Replaces objects in the storage.}
\usage{
save_objects(storagepath, obj.environment, objectnames = NULL,
  flag_forced_save_filenames = FALSE, flag_use_tmp_storage = FALSE,
  forced_archive_paths = NA, compress = "gzip",
  large_archive_prefix = NULL, locktimeout = NULL, wait_for = "save",
  parallel_cpus = NULL)
}
\arguments{
\item{storagepath}{Path with the storage.}

\item{obj.environment}{Environment or named list with the objects}

\item{objectnames}{Optional character vector with the names of the objects to add. Defaults to all objects in the
\code{obj.environment}.}

\item{flag_forced_save_filenames}{Optional boolean vector. Controls, whether force a particular object in its own dedicated archive.
Value can be either single boolean, or vector of booleans with the same size as
\code{objectnames}, or named boolean vector with keys values of \code{objectnames}. In the latter case,
non-mentioned objects will be assumed value \code{FALSE} (i.e. not forced filename).}

\item{compress}{Controls the compression of the archive. It is important to realize, that if the archive
had already contained some objects prior to modifying it and the modification would not remove the objects,
those objects will be re-compressed with the \code{compress} compression, since the archive will effectively
be re-added. Supports 3 calues: \code{none}, \code{gzip} and \code{xz}.
Value can be either single character, or vector of characters with the same size as
\code{objectnames}, or named character vector with keys values of \code{objectnames}. In the latter case,
non-mentioned objects will be assumed value \code{gzip}.}

\item{large_archive_prefix}{If set, all new archives for large objects will be saved with this prefix, otherwise in the
\code{dirname(storagepath)}.
\code{storagepath}. It is up to the user to make sure this directory is empty and no file name conflicts will
arise.}

\item{archive_filename}{Optional character vector with custom paths to the archives. Can be a single character object,
vector with the size of \code{objectnames} or named vector with keys from \code{objectnames}.}
}
\value{
Returns `data.frame` with the following columns:
\describe{
\item{\strong{objectname}}{Name of the stored object. This is a primary key.}
\item{\strong{digest}}{String with the digest of the object.}
\item{\strong{size}}{Numeric value with the size of the stored object.}
\item{\strong{archive_filename}}{Path where the object is stored absolute or relative to the storage path.}
\item{\strong{single_object}}{Logical. \code{TRUE} if the archive contain only this one object. Otherwise
archive contains named list of objects.}
\item{\strong{compress}}{Type of compression used to store this individual object}
}
}
\description{
Replaces objects in the storage.
}
