% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{set_runtime_objects}
\alias{set_runtime_objects}
\title{Sets new contents of the objectstorage.}
\usage{
set_runtime_objects(storagepath, obj.environment, objectnames = NULL,
  flag_forced_save_filenames = FALSE, flag_use_tmp_storage = FALSE,
  forced_archive_paths = NA, compress = "gzip",
  large_archive_prefix = NULL, locktimeout = NULL, wait_for = "save",
  parallel_cpus = NULL)
}
\arguments{
\item{storagepath}{Path with the storage.}

\item{obj.environment}{Environment or named list with the objects}

\item{objectnames}{Character vector with the names of the objects to add. Defaults to all objects in the
\code{obj.environment}.}

\item{flag_forced_save_filenames}{Boolean vector with the length equal to objectnames, or
named vector with objectnames as keys, or single value specifying whether to put a specific object
in its own dedicated archive}

\item{flag_use_tmp_storage}{Boolean vector with the length equal to objectnames, or
named vector with objectnames as keys, or single value specifying whether if the temporary
save file should be created in the fast /tmp directory first, and only then compressed into
the target place.}

\item{forced_archive_paths}{Character vector with the length equal to objectnames, or
named vector with objectnames as keys, or single value specifying custom path of archive
where a specific object(s) will saved.}

\item{compress}{Character vector with the length equal to objectnames, or
named vector with objectnames as keys, or single value specifying compression algorithm for
each object. Compression will be applied archive-wise.}

\item{large_archive_prefix}{If set, all new archives for large objects will be saved with this prefix, otherwise in the
\code{dirname(storagepath)}.
\code{storagepath}. It is up to the user to make sure this directory is empty and no file name conflicts will
arise.}
}
\value{
Returns `data.frame` with the following columns:
\describe{
\item{\strong{objectname}}{Name of the stored object. This is a primary key.}
\item{\strong{digest}}{String with the digest of the object.}
\item{\strong{size}}{Numeric value with the size of the stored object.}
\item{\strong{forced_archive_paths}}{Path where the object is stored absolute or relative to the storage path.}
\item{\strong{single_object}}{Logical. \code{TRUE} if the archive contain only this one object. Otherwise
archive contains named list of objects.}
\item{\strong{compress}}{Type of compression used to store this individual object}
}
}
\description{
The input objects will be compared with the stored objects, and replaced only when needed
}
