% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locks.R
\name{wait.for.lock}
\alias{wait.for.lock}
\title{Halts execution until the lock is released, or lock expires.}
\usage{
wait.for.lock(path, timeout = NULL)
}
\arguments{
\item{path}{Path to the lock file}

\item{timeout}{Wall time after which the lock will be considered open regardless if someone opened it}
}
\description{
Halts execution until the lock is released, or lock expires.
}
