% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{wait.for.save}
\alias{wait.for.save}
\title{Waits for .lock file to disappear}
\usage{
wait.for.save(path, name, timeout = 30 * 60)
}
\arguments{
\item{path}{Path to the lock file}

\item{name}{Name of the lock file. Both with \code{path} are used to get
the mangled connection name}

\item{timeout}{Timeout counted from file's creation time,
after which the lock will be removed automatically.}
}
\description{
Waits for .lock file to disappear
}
