% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectstorage.R
\name{load_objects}
\alias{load_objects}
\title{Loads specified objects into the given environment.}
\usage{
load_objects(storagepath, objectnames, target.environment,
  flag_double_check_digest = FALSE)
}
\arguments{
\item{storagepath}{Path to the storage}

\item{objectnames}{Vector with names of objects to load. Defaults to all objects}

\item{envir}{The target environment. Defaults to the global environment.}
}
\value{
Nothing (besides the \code{envir}). Environments get updated by reference.
}
\description{
Loads specified objects into the given environment.
}
